ccpd_alpr/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccpd-alpr"
version = "0.1.0"
description = "High-accuracy CCPD license plate localization + recognition pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.9",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "torch>=2.2",
  "torchvision>=0.17",
  "ultralytics>=8.3.0",
  "gradio>=4.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ccpd_alpr*"]
